feat-break-components: header, trending movies, movie list, useMovies hook
