+++ src/App.tsx
@@ -1,93 +1,25 @@
 import { useEffect, useState } from "react";
 import "./App.css";
-import Search from "./components/Search";
-import Spinner from "./components/Spinner";
-import MovieCard from "./components/MovieCard";
 import { useDebounce } from "react-use";
-import { getTrendingMovies, updateSearchCount } from "./libs/appwrite";
-
-const API_BASE_URL = "https://api.themoviedb.org/3";
-const API_KEY = import.meta.env.VITE_TMDB_API_KEY;
-
-const API_OPTIONS = {
-  method: "GET",
-  headers: {
-    accept: "application/json",
-    Authorization: `Bearer ${API_KEY}`,
-  },
-};
-
-interface Movie {
-  id: number;
-  title: string;
-  vote_average: number;
-  poster_path: string | null;
-  release_date: string;
-  original_language: string;
-}
-interface TrendingMovie {
-  $id: string;
-  title: string;
-  poster_url: string;
-}
+import Header from "./sections/Header";
+import TrendingMovies from "./components/TrendingMovies";
+import MovieList from "./components/MovieList";
+import { useMovies } from "./hooks/useMovies";
 
 const App = () => {
   const [searchTerm, setSearchTerm] = useState<string>("");
-  const [errorMessage, setErrorMessage] = useState<string>("");
-  const [movieList, setMovieList] = useState<Movie[]>([]);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-
   const [debouncedSearchTerm, setDebouncedSearchTerm] = useState("");
 
-  const [trendingMovies, setTrendingMovies] = useState<TrendingMovie[]>([]);
+  const {
+    movieList,
+    trendingMovies,
+    isLoading,
+    errorMessage,
+    fetchMovies,
+    loadTrendingMovies,
+  } = useMovies();
 
   useDebounce(() => setDebouncedSearchTerm(searchTerm), 500, [searchTerm]);
-
-  const fetchMovies = async (query = "") => {
-    setIsLoading(true);
-    setErrorMessage("");
-
-    try {
-      const endpoint = query
-        ? `${API_BASE_URL}/search/movie?query=${encodeURIComponent(query)}`
-        : `${API_BASE_URL}/discover/movie?sort_by=popularity.desc`;
-
-      const response = await fetch(endpoint, API_OPTIONS);
-
-      if (!response.ok) {
-        throw new Error("Failed to fetch movies");
-      }
-
-      const data = await response.json();
-
-      if (data.Response === "False") {
-        setErrorMessage(data.Error || "Failed to fetch movies");
-        setMovieList([]);
-        return;
-      }
-
-      setMovieList(data.results || []);
-
-      if (query && data.results.length > 0) {
-        await updateSearchCount(query, data.results[0]);
-      }
-    } catch (error) {
-      console.error(`Error fetching movies: ${error}`);
-      setErrorMessage("Error fetching movies. Please try again later.");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const loadTrendingMovies = async () => {
-    try {
-      const movies = await getTrendingMovies();
-
-      setTrendingMovies(movies);
-    } catch (error) {
-      console.error(`Error fetching trending movies: ${error}`);
-    }
-  };
 
   useEffect(() => {
     fetchMovies(debouncedSearchTerm);
@@ -102,45 +34,17 @@ const App = () => {
       <div className="pattern" />
 
       <div className="wrapper">
-        <header>
-          <img src="./hero.png" alt="Hero Banner" />
-          <h1>
-            Find <span className="text-gradient">Movies</span> You'll Enjoy
-            Without the Hassle
-          </h1>
-          <Search searchTerm={searchTerm} setSearchTerm={setSearchTerm} />
-        </header>
+        <Header searchTerm={searchTerm} setSearchTerm={setSearchTerm} />
 
         {trendingMovies.length > 0 && (
-          <section className="trending">
-            <h2>Trending Movies</h2>
-
-            <ul>
-              {trendingMovies.map((movie, index) => (
-                <li key={movie.$id}>
-                  <p>{index + 1}</p>
-                  <img src={movie.poster_url} alt={movie.title} />
-                </li>
-              ))}
-            </ul>
-          </section>
+          <TrendingMovies trendingMovies={trendingMovies} />
         )}
 
-        <section className="all-movies">
-          <h2>All Movies</h2>
-
-          {isLoading ? (
-            <Spinner />
-          ) : errorMessage ? (
-            <p className="text-red-500">{errorMessage}</p>
-          ) : (
-            <ul>
-              {movieList.map((movie) => (
-                <MovieCard key={movie.id} movie={movie} />
-              ))}
-            </ul>
-          )}
-        </section>
+        <MovieList
+          movieList={movieList}
+          isLoading={isLoading}
+          errorMessage={errorMessage}
+        />
       </div>
     </main>
   );
