+++ src/hooks/useMovies.ts
@@ -1,0 +1,89 @@
+import { useState } from "react";
+import { getTrendingMovies, updateSearchCount } from "../libs/appwrite";
+
+interface Movie {
+  id: number;
+  title: string;
+  vote_average: number;
+  poster_path: string | null;
+  release_date: string;
+  original_language: string;
+}
+
+interface TrendingMovie {
+  $id: string;
+  title: string;
+  poster_url: string;
+}
+
+const API_BASE_URL = "https://api.themoviedb.org/3";
+const API_KEY = import.meta.env.VITE_TMDB_API_KEY;
+
+const API_OPTIONS = {
+  method: "GET",
+  headers: {
+    accept: "application/json",
+    Authorization: `Bearer ${API_KEY}`,
+  },
+};
+
+export const useMovies = () => {
+  const [movieList, setMovieList] = useState<Movie[]>([]);
+  const [trendingMovies, setTrendingMovies] = useState<TrendingMovie[]>([]);
+  const [isLoading, setIsLoading] = useState(false);
+  const [errorMessage, setErrorMessage] = useState("");
+
+  const fetchMovies = async (query = "") => {
+    setIsLoading(true);
+    setErrorMessage("");
+
+    try {
+      const endpoint = query
+        ? `${API_BASE_URL}/search/movie?query=${encodeURIComponent(query)}`
+        : `${API_BASE_URL}/discover/movie?sort_by=popularity.desc`;
+
+      const response = await fetch(endpoint, API_OPTIONS);
+
+      if (!response.ok) {
+        throw new Error("Failed to fetch movies");
+      }
+
+      const data = await response.json();
+
+      if (data.Response === "False") {
+        setErrorMessage(data.Error || "Failed to fetch movies");
+        setMovieList([]);
+        return;
+      }
+
+      setMovieList(data.results || []);
+
+      if (query && data.results.length > 0) {
+        await updateSearchCount(query, data.results[0]);
+      }
+    } catch (error) {
+      console.error(`Error fetching movies: ${error}`);
+      setErrorMessage("Error fetching movies. Please try again later.");
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  const loadTrendingMovies = async () => {
+    try {
+      const movies = await getTrendingMovies();
+      setTrendingMovies(movies);
+    } catch (error) {
+      console.error(`Error fetching trending movies: ${error}`);
+    }
+  };
+
+  return {
+    movieList,
+    trendingMovies,
+    isLoading,
+    errorMessage,
+    fetchMovies,
+    loadTrendingMovies,
+  };
+};
